flatten tests, combining all API tests into one file
