+++ test/runtests.sh
@@ -1,7 +1,4 @@
 #!/bin/bash
 # USAGE: Run this file using `npm test` from repository root
 
-mocha \
-    test/*_test.js \
-    test/api/*_test.js \
-    ;
+mocha test/*_test.js
